client/desktop/requirements.txt
PyQt6>=6.6.0
markdown2>=2.4.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
Pillow>=10.0.0
requests>=2.28.0
